Mostly bugfixes and the testing of the functions, and yes they work, though need to be tweaked to be a bit more responsive to the player.

+++ Mario/MarioPOMDP/src/dk/itu/mario/MarioInterface/GamePlay.java
@@ -58,7 +58,7 @@ public class GamePlay implements Serializable {
     public int JumpFlowersKilled; //number of Jump Flower Mario killed
     public int ChompFlowersKilled; //number of Chomp Flower Mario killed
 
-    public int  totalRightTimeStraight,
+    public double  totalRightTimeStraight,
                 totalLeftTimeStraight,
                 totalRunTimeStraight,
                 totalRightTimeJump,
